[libcxx] Don't use __int128 for msvc targets
While clang-cl can compile it just fine, the compiler support lib
lacks support for it (as compiler-rt builtins aren't used).

This means that file_time_type has a shorter range on this target
(+/- 292 years from the epoch). This is smaller than the range of
the windows native file time type (which is 64 bit, with a 100 ns
unit, from an epoch of year 1601), but covers the practically used
range just fine. See docs/DesignDocs/FileTimeType.rst for more of
the tradeoffs this means.

Differential Revision: https://reviews.llvm.org/D91139