[libcxx] [test] Add XFAILs for mingw && windows-dll for overriding new/delete
When libc++ is built as a DLL, the weak operator new(nothrow) within
the DLL doesn't get overriden by the user executable's operator
new, so the fallback codepath from operator new(nothrow) to
operator new(throwing) calls the default operator new in the DLL
instead of the one in the user's executable.

This works as it should when using vcruntime as ABI library (i.e.
in MSVC/clang-cl configurations), because the nothrow operator new
is linked statically.