[OpenMP] [test] Don't use the compiler's omp.h in tasking/omp50_taskdep_depobj.c
When doing a all-in-one build of clang and openmp, this test
used the current omp.h, while if building openmp standalone,
it would use the compiler's omp.h, which only works correctly
if one already has installed a recent enough version of omp.h
into the compiler that was used.

Skip this discrepancy and just use the same headers as all other
tests.