WIP: [COFF] Remove the i386 underscore prefix from COFFImportFile::AliasTarget
On i386, regular C level symbols are given an underscore prefix
in the symbols on the object file level. However, the exported
names from DLLs usually don't have this leading underscore.

When specified in a def file like "symbol == dllname", the "dllname"
is the name of the exported symbol from the DLL, which will be
linked against from an object file symbol named "_symbol"
(on i386).

The mechanism where one symbol is redirected to another one in
an import library is implemented with weak aliases. In that case,
we need to have the object file symbol level name for the target
of the import, as we make one object file symbol point at another
one. Therefore, we added an underscore to the AliasTarget field.

This mechanism, with weak aliases, only works as long as the
target also is exported as is, in that form.

This alias mechanism detracts us from the original purpose of this
(slightly misnamed) field; it's the raw, undecorated name of
the DLL export symbol that we should link against.

Therefore, store the AliasTarget field unprefixed. When creating
the aliases, add the prefix as needed.