[libc++] Refactor the Windows and MinGW implementation of the locale base API
This patch reimplements the locale base support for Windows flavors
in a way that is more modules-friendly and without defining non-internal
names.

Since this changes the name of some types and entry points in the
built library, this is effectively an ABI break on Windows (which
is acceptable since we don't promise ABI stability on that platform).