[libcxx] [test] Remove unused alternate names for locales
If we detect that either of those alternative locale names can
be chosen, we set the locale.en_US.UTF-8 feature, which enables
tests that try to explicitly set and use the "en_US.UTF-8"
locale.

Thus, if a system only successfully can set a locale with the name
"English_United States.1252" (but not the others), this doesn't mean
that we can execute tests that tries enabling the "en_US.UTF-8"
locale.