[compiler-rt] Only consider asan supported on x86_32 and x86_64 on Windows
This avoids building asan when targeting Windows on armv7 or
aarch64. It is possible to build asan successfully for those
configurations (since 5ea9dd8c7076270695a1d90b9c73718e7d95e0bf
and 0c391133c9201ef29273554a1505ef855ce17668), but asan isn't
functional there.

This change skips building asan for targets other than x86_32
and x86_64.

By excluding asan from the build, we fix the "check-ubsan" target
for armv7 Windows. If asan is included in the build, an
ubsan-asan configuration gets added to the tests, and as asan
isn't functional for this target, it produces a lot of test failures
even when just trying to run "check-ubsan".

Curiously enough, this doesn't happen for aarch64 - since
00f3f6e296d49eb261e1ad47868a50122bfc111e in 2016, the ubsan-asan
tests are skipped on "64 bit Windows", which also ends up covering
aarch64. (We should probably revisit whether this is needed at
this point, or if we could remove this excemption, which seemed
to be meant to be temporary at the time.)